홈화면 및 Splash Screen 상세 구현(로딩중인 애니메이션) 완료

+++ lib/restaurant/model/product_response_dto_list.dart
@@ -5,7 +5,7 @@ part 'product_response_dto_list.g.dart';
 
 @JsonSerializable()
 class ProductResponseDtoList {
-  final int id;
+  final double id;
   final String productName;
 
   ProductResponseDtoList(this.id, this.productName);
